Revert "Promote KTO to stable API"
This reverts commit a4849786690bf238c7e41cf666b79c3079b3ecd8.
